accessible/tests/mochitest/test_aria_role_equation.html

<!DOCTYPE html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=428479
-->
<head>
  <title>ARIA role math chrome tests</title>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="common.js"></script>
  <script type="application/javascript"
          src="role.js"></script>

  <script type="application/javascript">
    function testThis(aID)
    {
      getRole(aID, ROLE_FLAT_EQUATION);
      var eqAcc = getAccessible(aID);
      if (eqAcc)
        is(eqAcc.name, "x^2 + y^2 + z^2", "Wrong name for " + aID + "!");
    }

    function doTest()
    {
      // Test equation image
      testThis("img_eq");

      // Test textual equation
      testThis("txt_eq");

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body>

  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=428479">Mozilla Bug 428479</a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>
  <p>Image: 
    <img id="img_eq" role="math" src="foo" alt="x^2 + y^2 + z^2">
  </p>
  <p>Text: 
    <span id="txt_eq" role="math" title="x^2 + y^2 + z^2">x<sup>2</sup> + 
      y<sup>2</sup> + z<sup>2</sup></span>
  </p>
</body>
</html>
